cvdef.hpp includes added

+++ modules/core/src/convert.cpp
@@ -302,7 +302,7 @@ static bool ocl_split( InputArray _m, OutputArrayOfArrays _mv )
         argidx = k.set(argidx, ocl::KernelArg::WriteOnlyNoSize(dst[i]));
     k.set(argidx, rowsPerWI);
 
-    size_t globalsize[2] = { size.width, (size.height + rowsPerWI - 1) / rowsPerWI };
+    size_t globalsize[2] = { static_cast<size_t>(size.width), static_cast<size_t>((size.height + rowsPerWI - 1) / rowsPerWI) };
     return k.run(2, globalsize, NULL, false);
 }
 
@@ -469,7 +469,7 @@ static bool ocl_merge( InputArrayOfArrays _mv, OutputArray _dst )
     argidx = k.set(argidx, ocl::KernelArg::WriteOnly(dst));
     k.set(argidx, rowsPerWI);
 
-    size_t globalsize[2] = { dst.cols, (dst.rows + rowsPerWI - 1) / rowsPerWI };
+    size_t globalsize[2] = { static_cast<size_t>(dst.cols), static_cast<size_t>((dst.rows + rowsPerWI - 1) / rowsPerWI) };
     return k.run(2, globalsize, NULL, false);
 }
 
@@ -738,7 +738,7 @@ static bool ocl_mixChannels(InputArrayOfArrays _src, InputOutputArrayOfArrays _d
     argindex = k.set(argindex, size.width);
     k.set(argindex, rowsPerWI);
 
-    size_t globalsize[2] = { size.width, (size.height + rowsPerWI - 1) / rowsPerWI };
+    size_t globalsize[2] = { static_cast<size_t>(size.width), static_cast<size_t>((size.height + rowsPerWI - 1) / rowsPerWI) };
     return k.run(2, globalsize, NULL, false);
 }
 
@@ -1572,7 +1572,7 @@ static bool ocl_convertScaleAbs( InputArray _src, OutputArray _dst, double alpha
     else if (wdepth == CV_64F)
         k.args(srcarg, dstarg, alpha, beta);
 
-    size_t globalsize[2] = { src.cols * cn / kercn, (src.rows + rowsPerWI - 1) / rowsPerWI };
+    size_t globalsize[2] = { static_cast<size_t>(src.cols * cn / kercn), static_cast<size_t>((src.rows + rowsPerWI - 1) / rowsPerWI) };
     return k.run(2, globalsize, NULL, false);
 }
 
@@ -1744,7 +1744,7 @@ static bool ocl_LUT(InputArray _src, InputArray _lut, OutputArray _dst)
     k.args(ocl::KernelArg::ReadOnlyNoSize(src), ocl::KernelArg::ReadOnlyNoSize(lut),
         ocl::KernelArg::WriteOnlyNoSize(dst), dst.rows, cols);
 
-    size_t globalSize[2] = { cols, (dst.rows + 3) / 4 };
+    size_t globalSize[2] = { static_cast<size_t>(cols), static_cast<size_t>((dst.rows + 3) / 4) };
     return k.run(2, globalSize, NULL, false);
 }
 
@@ -2089,7 +2089,7 @@ static bool ocl_normalize( InputArray _src, InputOutputArray _dst, InputArray _m
                 k.args(srcarg, maskarg, dstarg);
         }
 
-        size_t globalsize[2] = { src.cols, (src.rows + rowsPerWI - 1) / rowsPerWI };
+        size_t globalsize[2] = { static_cast<size_t>(src.cols), static_cast<size_t>((src.rows + rowsPerWI - 1) / rowsPerWI) };
         return k.run(2, globalsize, NULL, false);
     }
     else
